+++ src/commands/index.js
@@ -1,6 +1,5 @@
 import {CommandSet} from 'prosemirror/src/edit'
 import edCommands from './ed-commands'
-import {joinBackward} from './schema-commands'
 
 let commands = CommandSet.default
   .add(edCommands)
@@ -8,7 +7,6 @@ let commands = CommandSet.default
     { 'code:toggle': {menu: null}
     , joinUp: {menu: null}
     , selectParentNode: {menu: null}
-    , joinBackward
     }
   )