+++ test/convert/doc-to-grid.js
@@ -9,17 +9,13 @@ import EdSchemaFull from '../../src/schema/ed-schema-full'
 describe('DocToGrid', function () {
   describe('with full schema', function () {
     const content =
-      [ {type: 'h1', html: '<h1>heading 1</h1>'}
-      , {type: 'h2', html: '<h2>heading 2</h2>'}
-      , {type: 'h3', html: '<h3>heading 3</h3>'}
-      , {type: 'text', html: '<p>paragraph 1</p>'}
-      , { id: 'image-0000'
-        , type: 'image'
-        , cover: {src: '..jpg'}
-        , html: '<img src="..jpg">'
-        }
-      , {id: 'video-0000', type: 'video'}
-      , {type: 'quote', html: '<blockquote><p>bq</p></blockquote>'}
+      [ {type: 'h1', html: '<h1>heading 1</h1>', metadata: {starred: true}}
+      , {id: 'image-0000', type: 'image', metadata: {starred: true}}
+      , {type: 'text', html: '<p>paragraph 1</p>', metadata: {starred: true}}
+      , {type: 'h2', html: '<h2>heading 2</h2>', metadata: { starred: false }}
+      , {type: 'h3', html: '<h3>heading 3</h3>', metadata: { starred: false }}
+      , {id: 'video-0000', type: 'video', metadata: {starred: false}}
+      , {type: 'quote', html: '<blockquote><p>bq</p></blockquote>', metadata: { starred: false }}
       ]
     const map = {}
     for (let i = 0, len = content.length; i < len; i++) {
@@ -31,10 +27,21 @@ describe('DocToGrid', function () {
     const doc =
       { 'type': 'doc'
       , 'content':
-      [
-          { 'type': 'heading'
+        [ { 'type': 'heading'
           , 'attrs': {'level': 1}
           , 'content': [{'type': 'text', 'text': 'heading 1'}]
+          }
+        , { 'type': 'media'
+          , 'attrs':
+            { 'id': 'image-0000'
+            , 'type': 'image'
+            , 'height': 50
+            }
+          }
+        , { 'type': 'paragraph'
+          , 'content': [{'type': 'text', 'text': 'paragraph 1'}]
+          }
+        , { 'type': 'horizontal_rule'
           }
         , { 'type': 'heading'
           , 'attrs': {'level': 2}
@@ -43,16 +50,6 @@ describe('DocToGrid', function () {
         , { 'type': 'heading'
           , 'attrs': {'level': 3}
           , 'content': [{'type': 'text', 'text': 'heading 3'}]
-          }
-        , { 'type': 'paragraph'
-          , 'content': [{'type': 'text', 'text': 'paragraph 1'}]
-          }
-        , { 'type': 'media'
-          , 'attrs':
-            { 'id': 'image-0000'
-            , 'type': 'image'
-            , 'height': 50
-            }
           }
         , { 'type': 'media'
           , 'attrs':
@@ -63,13 +60,12 @@ describe('DocToGrid', function () {
           }
         , { 'type': 'blockquote'
           , 'content':
-          [
-              { 'type': 'paragraph'
+            [ { 'type': 'paragraph'
               , 'content': [{'type': 'text', 'text': 'bq'}]
               }
-          ]
+            ]
           }
-      ]
+        ]
       }
 
     it('correctly converts full Doc to Grid content', function () {