+++ test/index.js
@@ -2,7 +2,6 @@ import chai from 'chai'
 chai.config.truncateThreshold = 0
 
 import './ed'
-import './components/fold-media'
 import './components/image'
 import './convert/determine-fold'
 import './convert/doc-to-grid.js'