+++ test/plugins/widget.js
@@ -6,16 +6,16 @@ import Ed from '../../src/ed'
 describe('PluginWidget', function () {
   let mount, ed, PluginWidget
   const fixture =
-    [ {type: 'h1', html: '<h1>Title</h1>'}
+    [ {type: 'h1', html: '<h1>Title</h1>', metadata: {starred: true}}
     , { id: '0001'
       , type: 'placeholder'
-      , metadata: {status: 'Status'}
+      , metadata: {status: 'Status', starred: true}
       }
     , { id: '0000'
       , type: 'placeholder'
-      , metadata: {status: 'Status'}
+      , metadata: {status: 'Status', starred: true}
       }
-    , {type: 'text', html: '<p>Text</p>'}
+    , {type: 'text', html: '<p>Text</p>', metadata: {starred: true}}
     ]
 
   beforeEach(function (done) {
@@ -39,17 +39,20 @@ describe('PluginWidget', function () {
 
   describe('Content mounting and merging', function () {
     it('has expected pm document', function () {
-      expect(ed._foldMedia).to.equal('0001')
       const content = ed.pm.doc.content.content
-      expect(content.length).to.equal(3)
+      expect(content.length).to.equal(4)
       expect(content[0].textContent).to.equal('Title')
       expect(content[0].type.name).to.equal('heading')
       expect(content[1].textContent).to.equal('')
       expect(content[1].type.name).to.equal('media')
-      expect(content[1].attrs.id).to.equal('0000')
+      expect(content[1].attrs.id).to.equal('0001')
       expect(content[1].attrs.type).to.equal('placeholder')
-      expect(content[2].textContent).to.equal('Text')
-      expect(content[2].type.name).to.equal('paragraph')
+      expect(content[2].textContent).to.equal('')
+      expect(content[2].type.name).to.equal('media')
+      expect(content[2].attrs.id).to.equal('0000')
+      expect(content[2].attrs.type).to.equal('placeholder')
+      expect(content[3].textContent).to.equal('Text')
+      expect(content[3].type.name).to.equal('paragraph')
     })
 
     it('has mounted widget', function () {
@@ -135,10 +138,10 @@ describe('PluginWidget', function () {
 
       // PM placeholder change is sync
       const content = ed.pm.doc.content.content
-      expect(content[1].textContent).to.equal('')
-      expect(content[1].type.name).to.equal('media')
-      expect(content[1].attrs.id).to.equal('0000')
-      expect(content[1].attrs.type).to.equal('image')
+      expect(content[2].textContent).to.equal('')
+      expect(content[2].type.name).to.equal('media')
+      expect(content[2].attrs.id).to.equal('0000')
+      expect(content[2].attrs.type).to.equal('image')
     })
   })
 })